Improved examples, created a new file for generating sample data.'
